updated tests to reflect new structure (utils folder)
