+++ tests/test_utils/test_db.py
@@ -1,0 +1,12 @@
+import unittest
+from utils import u_db
+
+class UtilsTestCase(unittest.TestCase):
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    def test_get_db_conn_and_cursor(self):
+        pass
